requirements.txt
packaging==25.0
pdf2image==1.17.0
pillow==10.4.0
pytesseract==0.3.10
